rename tables with the same names
